+++ lib/toolbar/models/header-styles.const.dart
@@ -1,0 +1,10 @@
+// Default header styles used in the editor
+// Every style is linked to the suitable attribute
+const Map<String, int> HEADER_STYLES = {
+  'N': 0,
+  'H1': 1,
+  'H2': 2,
+  'H3': 3,
+};
+
+const INITIAL_HEADER_STYLE = 0;